requirements.txt
jina>=3.9.1
kafka-python>=2.0.2